requirements.txt
# Sphery Motivational Engine - Dependencies
# Core ML and Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
mysql-connector-python>=8.0.0

# Dashboard and Web Interface
streamlit>=1.28.0

# Model Persistence
joblib>=1.3.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.0.0

# Development and Testing
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Additional Utilities
python-dateutil>=2.8.0 